requirements.txt
python-dotenv
openai
spacy
numpy